Conditional fixes, and linting
When checking the value of variable in a conditional, checks should
not use the "is" keyword. Boolean checks should be in the form of
"if VARIABLE". Due to object caching in the interpreter, compairing
values with "is" can fail as you're checking for reference equality
which can change as objects drop out of the cache.

https://stackoverflow.com/questions/132988/is-there-a-difference-between-and-is-in-python